Add tests to load sample config
