+++ __tests__/load-config.test.ts
@@ -1,4 +1,4 @@
-import { loadConfig } from '../src/config'
+import { loadConfig } from '../src/load-config'
 import * as fs from 'fs'
 import * as yaml from 'js-yaml'
